requirements.txt
flask
google-generativeai
pandas
pypdf
python-dotenv
fpdf
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
